\name{na.as.zero}
\alias{na.as.zero}
\title{na as zero}
\description{Convert empty string to zero and use as.integer otherwise.}
\usage{na.as.zero(int.or.empty)}
\arguments{
  \item{int.or.empty}{int.or.empty }
}

\value{Integer vector.}

\author{Toby Hocking <toby.hocking@r-project.org> [aut, cre]}





