\name{sacct-2019-11-21}
\alias{sacct-2019-11-21}
\docType{data}
\title{
  Output of sacct for thousands of jobs
}
\description{
  The data come from running sacct on a SLURM system.
}
\format{
This "csv" file results from sacct -P which uses the vertical bar | as
the separator. There are five columns and six rows (including header).
}
\keyword{datasets}
