\name{sacct-job13936577}
\alias{sacct-job13936577}
\docType{data}
\title{
  Output of sacct for one job with several tasks
}
\description{
  The data come from running sacct on a SLURM system.
}
\format{
This "csv" file results from sacct -P which uses the vertical bar | as
the separator. There are five columns and 709 rows (including header).
}
\keyword{datasets}
