\name{sacct-multitype}
\alias{sacct-multitype}
\docType{data}
\title{
  Output of sacct for three job with several tasks
}
\description{
  The data come from running sacct on a SLURM system.
}
\format{
This "csv" file results from sacct -P which uses the vertical bar | as
the separator. There are five columns and six rows (including header).
}
\keyword{datasets}
