\name{sjob}
\alias{sjob}
\title{sjob}
\description{Run \code{\link{sacct}} and summarize State/ExitCode values for given job IDS}
\usage{sjob(job.id = sq.jobs(), 
    tasks.width = 11, 
    ...)}
\arguments{
  \item{job.id}{job.id }
  \item{tasks.width}{tasks.width }
  \item{\dots}{\dots }
}

\value{Data table from \code{\link{sjob_dt}}.}

\author{Toby Hocking <toby.hocking@r-project.org> [aut, cre]}





