\name{sjob}
\alias{sjob}
\title{sjob}
\description{Report counts of tasks with State/ExitCode values for given job IDS}
\usage{sjob(job.id = sq.jobs(), tasks.width = 11)}
\arguments{
  \item{job.id}{
}
  \item{tasks.width}{
}
}

\value{data.table summarizing State/ExitCode distribution over jobs}

\author{Toby Dylan Hocking}





